Refactor api client access for configurable cache
