+++ lib/stitches/configuration.rb
@@ -13,8 +13,8 @@ class Stitches::Configuration
     @custom_http_auth_scheme = UnsetString.new("custom_http_auth_scheme")
     @env_var_to_hold_api_client_primary_key = NonNullString.new("env_var_to_hold_api_client_primary_key","STITCHES_API_CLIENT_ID")
     @env_var_to_hold_api_client= NonNullString.new("env_var_to_hold_api_client","STITCHES_API_CLIENT")
-    @max_cache_ttl = NonNullInteger.new("max_cache_ttl", 5 * 60)
-    @max_cache_size = NonNullInteger.new("max_cache_size", 100)
+    @max_cache_ttl = NonNullInteger.new("max_cache_ttl", 0)
+    @max_cache_size = NonNullInteger.new("max_cache_size", 0)
   end
 
   # A RegExp that allows URLS around the mime type and api key requirements.
