+++ lib/stitches/railtie.rb
@@ -1,6 +1,6 @@
 require 'stitches/api_key'
 require 'stitches/valid_mime_type'
-require 'stitches/api_key_cache_wrapper'
+require 'stitches/api_client_access_wrapper'
 
 module Stitches
   class Railtie < Rails::Railtie
